animal movement restrict to grid

+++ src/Utility/Vec2.cpp
@@ -1,7 +1,7 @@
 /**
  * @file Vec2.cpp
  * @brief [Description of file]
- * 
+ *
  * @author Filip Dabkowski
  * @date 20/03/2024
  */
@@ -10,14 +10,10 @@
 #include <cmath>
 
 
-float Vec2::dot(Vec2 vec1, Vec2 vec2) {
-	return vec1.x * vec2.x + vec1.y * vec2.y;
-}
+float Vec2::dot(Vec2 vec1, Vec2 vec2) { return vec1.x * vec2.x + vec1.y * vec2.y; }
 
 
-float Vec2::dot(Vec2 vec) {
-	return Vec2::dot(*this, vec);
-}
+float Vec2::dot(Vec2 vec) { return Vec2::dot(*this, vec); }
 
 
 float Vec2::magnitude2() {
@@ -27,13 +23,11 @@ float Vec2::magnitude2() {
 
 
 float Vec2::length(Vec2 vec1, Vec2 vec2) {
-	return sqrtf((vec1.x - vec2.x) * (vec1.x - vec2.x) + (vec1.y - vec2.y) * (vec1.y - vec2.y));
+	return sqrtf((float) ((vec1.x - vec2.x) * (vec1.x - vec2.x) + (vec1.y - vec2.y) * (vec1.y - vec2.y)));
 }
 
 
-float Vec2::length() {
-	return sqrtf(magnitude2());
-}
+float Vec2::length() { return sqrtf(magnitude2()); }
 
 
 Vec2 Vec2::normalised() {
@@ -53,11 +47,7 @@ Vec2 Vec2::scalarMultiply(float scalar) {
 }
 
 
-Vec2 Vec2::unitX() {
-	return {1, 0};
-}
+Vec2 Vec2::unitX() { return {1, 0}; }
 
 
-Vec2 Vec2::unitY() {
-	return {0, 1};
-}
+Vec2 Vec2::unitY() { return {0, 1}; }
